server/requirements.txt
2to3==1.0
absl-py==0.9.0
aniso8601==8.1.0
astor==0.8.1
certifi==2018.8.24
chardet==4.0.0
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
Flask-RESTful==0.3.8
gast==0.2.2
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==1.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.0.3
networkx==2.4
numpy==1.14.5
opencv-python==3.2.0.6
opt-einsum==3.2.1
Pillow==7.1.2
protobuf==3.12.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
requests==2.25.1
scikit-image==0.15.0
scikit-learn==0.22.2.post1
scipy==1.0.0
selenium==3.141.0
six==1.15.0
sklearn==0.0
tensorboard==1.10.0
tensorflow==1.10.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
urllib3==1.26.2
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==1.2.0
